requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
pandas==2.2.0
sqlcipher3-binary==0.5.2
joblib==1.3.2
numpy==1.26.4
xgboost==2.0.3
pytest==8.0.0
httpx==0.26.0
